Add request header strategies to allow extensibility
The default strategy behaves the same way as the original behaviour, which
is to return the headers from the route rules.

Making this into a Strategy means that projects may add their own strategies
to control which headers are shown in various circumstances.

This commit also includes some tidying up of the example request templates
for cases when no headers are required for the request.